Move merge util function to boss.core.util.object
